separate opcode execution logic from chip8
Now that State is a more abstract concept and opcode execution was
already returning a new one every time it seems sensible to make the
Chip8 execute_op a method on Opcodes. This is an attempt to narrow the
scope of Chip8's responsibility to just interfacing with the outside
world and tracking state.